add subscription mutation and fix error handling

+++ src/graphql/mutations.js
@@ -1,7 +1,7 @@
 import {gql} from 'apollo-boost'
 
 export const ADD_SONG = gql `
-mutation addSong($title: String!, $artist: String!, $url: String!, $thumbnail: String!, $duration: Float!) {
+mutation addSong($title: String!, $artist: String!, $url: String!, $thumbnail: String!, $duration: Int!) {
     insert_songs(objects: {thumbnail: $thumbnail, url: $url, artist: $artist, title: $title, duration: $duration}) {
       affected_rows
     }
